refactor: package equip together
