+++ HospitalSystem.Core/Equipment/EquipmentBatchRepository.cs
@@ -1,7 +1,7 @@
 using MongoDB.Driver;
 using System.Linq.Expressions;
 
-namespace HospitalSystem.Core;
+namespace HospitalSystem.Core.Equipment;
 
 public class EquipmentBatchRepository : IEquipmentBatchRepository
 {
