+++ HospitalSystem.Core/Equipment/EquipmentBatchService.cs
@@ -1,6 +1,6 @@
 using HospitalSystem.Core.Rooms;
 
-namespace HospitalSystem.Core;
+namespace HospitalSystem.Core.Equipment;
 
 [System.Serializable]
 public class EquipmentBatchDoesNotExistException : System.Exception
