+++ HospitalSystem.Core/Equipment/Orders/EquipmentOrderRepository.cs
@@ -1,6 +1,6 @@
 using MongoDB.Driver;
 
-namespace HospitalSystem.Core;
+namespace HospitalSystem.Core.Equipment;
 
 public class EquipmentOrderRepository : IEquipmentOrderRepository
 {
