+++ HospitalSystem.Core/Equipment/Orders/EquipmentOrderService.cs
@@ -1,6 +1,6 @@
 using HospitalSystem.Core.Utils;
 
-namespace HospitalSystem.Core;
+namespace HospitalSystem.Core.Equipment;
 
 public class EquipmentOrderService
 {
