+++ HospitalSystem.Core/Equipment/Relocations/EquipmentRelocationRepository.cs
@@ -1,6 +1,6 @@
 using MongoDB.Driver;
 
-namespace HospitalSystem.Core;
+namespace HospitalSystem.Core.Equipment;
 
 public class EquipmentRelocationRepository : IEquipmentRelocationRepository
 {
