+++ HospitalSystem.Core/Renovations/RenovationService.cs
@@ -1,5 +1,6 @@
 using HospitalSystem.Core.Utils;
 using HospitalSystem.Core.Rooms;
+using HospitalSystem.Core.Equipment;
 
 namespace HospitalSystem.Core.Renovations;
