+++ Tests/TestGenerator.cs
@@ -3,6 +3,7 @@ using MongoDB.Bson.IO;
 using MongoDB.Driver;
 using HospitalSystem.Core;
 using HospitalSystem.Core.Rooms;
+using HospitalSystem.Core.Equipment;
 using HospitalSystem.Core.Surveys;
 using HospitalSystem.Core.Utils;
